feat(Appointment): Implement UnauthorizedAppointmentAccessException and restrict appointment deletion to authorized users

+++ src/main/java/bodycheck_back/bodycheck/services/AppointmentService.java
@@ -2,7 +2,6 @@ package bodycheck_back.bodycheck.services;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
@@ -12,6 +11,7 @@ import bodycheck_back.bodycheck.auth.services.AuthService;
 import bodycheck_back.bodycheck.exceptions.appointment.AppointmentConflictException;
 import bodycheck_back.bodycheck.exceptions.appointment.AppointmentCustomerExpectedException;
 import bodycheck_back.bodycheck.exceptions.appointment.AppointmentNotFoundException;
+import bodycheck_back.bodycheck.exceptions.appointment.UnauthorizedAppointmentAccessException;
 import bodycheck_back.bodycheck.models.dtos.AppointmentDTO;
 import bodycheck_back.bodycheck.models.entities.Appointment;
 import bodycheck_back.bodycheck.models.entities.Customer;
@@ -25,12 +25,6 @@ public class AppointmentService {
    private final AppointmentRepository appointmentRepository;
    private final AuthService authService;
    private final CustomerService customerService;
-
-   public AppointmentDTO findById(Long id) {
-      Optional<Appointment> o = appointmentRepository.findById(id);
-
-      return convertToDto(o.orElseThrow(AppointmentNotFoundException::new));
-   }
 
    public List<AppointmentDTO> findAllByDate(LocalDate date) {
       List<Appointment> appointments = appointmentRepository.findAllByUserAndDate(authService.getUserFromToken(), date);
@@ -48,12 +42,15 @@ public class AppointmentService {
    @Transactional
    public AppointmentDTO create(Appointment appointment) {
       appointment.setUser(authService.getUserFromToken());
-      // IMPORTANTE! sumo 1s al startTime para que no encuentre conflicto cuando una cita empieza justo cuando otra acaba.
+      // IMPORTANTE! sumo 1s al startTime para que no encuentre conflicto cuando una
+      // cita empieza justo cuando otra acaba.
       appointment.setStartTime(appointment.getStartTime().plusSeconds(1));
 
-      // Comprobamos que el appointment este relacionado a un Customer o un Futuro Customer.
+      // Comprobamos que el appointment este relacionado a un Customer o un Futuro
+      // Customer.
       if ((appointment.getCustomer() == null || appointment.getCustomer().getId() == null)
-            && (appointment.getCustomerName() == null || appointment.getCustomerPhone() == null || appointment.getCustomerName() == "" || appointment.getCustomerPhone() == "")) {
+            && (appointment.getCustomerName() == null || appointment.getCustomerPhone() == null
+                  || appointment.getCustomerName() == "" || appointment.getCustomerPhone() == "")) {
          throw new AppointmentCustomerExpectedException();
       }
 
@@ -70,12 +67,15 @@ public class AppointmentService {
    public AppointmentDTO update(Long id, Appointment appointment) {
       appointment.setId(id);
       appointment.setUser(authService.getUserFromToken());
-      // IMPORTANTE! sumo 1s al startTime para que no encuentre conflicto cuando una cita empieza justo cuando otra acaba.
+      // IMPORTANTE! sumo 1s al startTime para que no encuentre conflicto cuando una
+      // cita empieza justo cuando otra acaba.
       appointment.setStartTime(appointment.getStartTime().plusSeconds(1));
 
-      // Comprobamos que el appointment este relacionado a un Customer o un Futuro Customer.
+      // Comprobamos que el appointment este relacionado a un Customer o un Futuro
+      // Customer.
       if ((appointment.getCustomer() == null || appointment.getCustomer().getId() == null)
-            && (appointment.getCustomerName() == null || appointment.getCustomerPhone() == null || appointment.getCustomerName() == "" || appointment.getCustomerPhone() == "")) {
+            && (appointment.getCustomerName() == null || appointment.getCustomerPhone() == null
+                  || appointment.getCustomerName() == "" || appointment.getCustomerPhone() == "")) {
          throw new AppointmentCustomerExpectedException();
       }
 
@@ -91,11 +91,22 @@ public class AppointmentService {
 
    @Transactional
    public void delete(Long id) {
-      appointmentRepository.deleteById(id);
+      Appointment appointment = findById(id);
+
+      if (appointment.getUser().getId().equals(authService.getUserFromToken().getId())) {
+         appointmentRepository.deleteById(id);
+      } else {
+         throw new UnauthorizedAppointmentAccessException();
+      }
+   }
+
+   private Appointment findById(Long id) {
+      return appointmentRepository.findById(id).orElseThrow(AppointmentNotFoundException::new);
    }
 
    public AppointmentDTO convertToDto(Appointment appointment) {
-      // Accedemos al Customer relaccionado con el apointment si existe. Si no, lo mantenemos como null.
+      // Accedemos al Customer relaccionado con el apointment si existe. Si no, lo
+      // mantenemos como null.
       Customer customer = appointment.getCustomer() != null
             ? customerService.findById(appointment.getCustomer().getId()).orElseThrow()
             : null;
